Use the Transfer Objects to share data (for User)
for the convenience of data exchange between the frontend and the server, the Transfer Objects object (and pattern) is used. The TO object contains only those fields that are needed by the UI, and in the process of running the application, the entity is converted to TO and vice versa.